Refactoring: move all rxjs subjects to separate file
