+++ src/components/Controls.tsx
@@ -1,14 +1,13 @@
 import { Dispatch, SetStateAction, useContext, useEffect, useState } from "react";
-import { assistant, TasksMatrix } from "./Assistant/assistant.lib";
 import { AuthData, EMCategory, Task, TaskType } from "../lib/types";
 import { Button } from "./ui/Button";
-import { convertServerDataToLocalData, filterLocalDataByCategory, tasksQ1, tasksQ2, tasksQ3, tasksQ4, tasksUncategorized, toDosQuery, useTodoTasksMutation } from "@/components/TaskLabel/lib/tasks.fn";
-import { moveTask, TasksListAction } from "./containers/dnd.lib";
+import { convertServerDataToLocalData, filterLocalDataByCategory, toDosQuery, useTodoTasksMutation } from "@/components/TaskLabel/lib/tasks.fn";
 import { CheckboxWL } from "./ui/CheckboxWL";
 import { AuthContext } from "@/lib/contexts";
 import { useStorageState } from "@/lib/hooks/use-storage-state";
 import { useToast } from "@/lib/hooks/use-toast";
 import { AssistantContainer } from "./Assistant/AssistantContainer";
+import { assistant, moveTask, TasksListAction, TasksMatrix, tasksQ1, tasksQ2, tasksQ3, tasksQ4, tasksUncategorized } from "@/lib/subjects";
 
 interface Quadrants {
   get: Task[],
