+++ src/components/containers/SidebarTasksList.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { EMCategory, Task } from "../../lib/types";
 import { TaskLabel } from "../TaskLabel/TaskLabel";
 import { useDroppable } from "@dnd-kit/core";
-import { getTasksFactory } from "../TaskLabel/lib/tasks.fn"
+import { getTasksFactory } from "@/lib/subjects";
 
 export function SidebarTasksList() {
   const [ tasks, setTasks ] = useState<Task[]>([]);
